Reorganize utils into specific directories
Trying to get the root repo into fewer folders and move files in the
root into directories.
