+++ demos/factor_compare.sh
