+++ factorization_util/VISUALS.md
