+++ factorization_util/view_npy_as_heatmap.py
